requirements.txt
apify-client>=1.7.0
google-genai>=0.3.0
python-dotenv>=1.0.0
